requirements.txt
# Core web and UI
streamlit==1.39.0
flask==3.0.3
flask-cors==4.0.1

# NLP & ML (lightweight)
textstat==0.7.4

# Parsing (pure Python)
pdfminer.six==20240706
docx2txt==0.8

# Data & DB
SQLAlchemy==2.0.35

# Utils
requests==2.32.3
reportlab==4.2.5